[json] Use map cache for stringify fast-path
We expect that many objects will share the same map, so we can cache
some interesting properties per map that we have fully serialized.
Namely:
- Are all properties enumerable.
- No property key is a symbol.
- No property key contains characters that require escaping.

When serializing an object with a map known to fulfill these
properties, we can omit a bunch of checks.

Bug: 380044242
Change-Id: I3f1975ba0928271df086717b6681cb9ec1ae97bd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6396098
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99525}