[js_fuzzer] Implement more sloppy execution order for available variables
This adds a workaround to improve available variables after the
@babel/traverse update. After a fix with the update from
https://crrev.com/c/6383037 there are fewer available variables
within functions, because the execution-order guess got more
conservative.

This improves this situation for function expressions and mitigates
it for function declarations behind a probability with which the
conservative choice is bypassed. This alternative choice, however,
does, when chosen, again simulate the behavior of before, which
can also cause rare reference errors, when function calls are moved
before the variable declarations.

Patch 9 adds the implementation and new expressive tests, which
include function expressions. The implementation is not chosen yet
to show the status quo:
https://crrev.com/c/6401114/9

Patch 10 turns on the new logic and shows the test changes. The
probability behind a flag is shown in two tests, once always chosen
and once never. Some pseudo-random tests show various random and
unrelated changes as often, since now an early random decision
is added whenever random variables are chosen:
https://crrev.com/c/6401114/9..10

BYPASS_LARGE_CHANGE_WARNING

Bug: 389069288
Change-Id: I03407877d98d8c563e6b30a9be3f67d1e719411e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6401114
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99569}