[json] Fix stringifier for non-fast map case
With the introduction of the fast map cache, we directly use the
descriptor number as the descriptor index. This only valid for the fast
map case (as we know that all properties are enumerable and not stored
in the descriptor), but it was accidentally also used while building
the map cache.

Fixed: 406828712
Bug: 380044242
Change-Id: I6d2fb867008c4496f74a7943d04707ee80d58df6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6404677
Auto-Submit: Patrick Thier <pthier@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99541}