[wasm][jspi] Do not process stack switches eagerly in unwinder
When an exception crosses a stack boundary, do not process the implicit
stack switch immediately, wait until we have found the handler.
Otherwise, the stack switching state is temporarily out of sync with the
thread local top, which is visible and can cause a crash if the
exception causes a deopt deeper in the stack.

R=jkummerow@chromium.org

Fixed: 406053619
Change-Id: Ib1e97227b3fdd0321ba86eec0f2eb856eda97e1f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6404813
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99542}