[json] Allow stringify slow-path directly after encoding change
Prior to [1] it was guaranteed that serializing a simple property after
an encoding change triggered by an object key will succeed, as it was
already checked that the object value can be serialized on the fast-path
before serializing the object key.
With that change, we don't check if the property value can be serialized
on the fast-path before serializing the key anymore, thus we can
fall-back to the slow-path immediately after resuming due to an encoding
change.

[1] https://crrev.com/c/6308767

Fixed: 406830031
Bug: 380044242
Change-Id: Ic20be06f2f9653f670d0b169ccddea9ae6f6c2b7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6405542
Commit-Queue: Patrick Thier <pthier@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Auto-Submit: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99543}