[wasm] Remove an overly strict DCHECK
Remove an overly strict DCHECK which helped to catch a few issues where
we continued execution after an exception (including termination
exceptions), but is unfortunately incompatible with first setting a
return value (e.g. a promise) and then still executing some JS code
which might trigger an exception.
Fixing all those places to clear the return value on exception is not
worth the hassle.

In order to protect against accidentally passing the `exception` or
`termination_exception` value to JS, we made the types more strict in
https://crrev.com/c/6375484.

R=jkummerow@chromium.org

Fixed: 403308742
Change-Id: I3e7b90218756bb24fce346a9239b9d37423e704a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6402514
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99546}