[heap] Limit target_capacity in FinishShrinking()
After MinorMS the current capacity could be above the maximum capacity
if not enough pages either die or get promoted. This in turns would
let the target capacity go above the maximum capacity. This
triggered DCHECKs and so this CL ensures that we stay within the
limits for target capacity.

Bug: 407817244
Change-Id: I09f999896d3d5d05ec62fef2680eba4fc2b5bc7f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6425340
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99637}