[heap] Do not lock mutex in PersistentHandlesList::Iterate()
This method can only be used in the GC pause and therefore there is
no need to lock the mutex in the first place.

This mutex is involved in a potential deadlock reported by
Clusterfuzz in the linked bug. The issue here is that for freeing
a TurbofanCompilationJob we also need to remove the PersistentHandles
from PersistentHandlesList.

Bug: 407609361
Change-Id: I73fec0422cceb1cad9c088747d9f80754807aff4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6424463
Auto-Submit: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99642}