[js_fuzzer] Make DB index scale better
This streamlines the DB index and bootstraps it at fuzzer start-up,
which takes below a millisecond.

The streamlined version makes adding more tags for each stored
expression scale. E.e. in a follow up we want to tag expressions
to be chosen at a lower probability.

This also adds formatting to the JSON to make it more readible.

With this update, the DB will need to be rebuilt and all fuzzer
instances need to be updated with the new version.

Bug: 389069288
Change-Id: I46808fd80959342d72f26f495c31009747f02959
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6429848
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99654}