[presubmit] Fix presubmit check for inline headers on Windows
One of the 'git cl upload' presubmit checks is that the first include
of an -inl.h header should be the non-inl counterpart. This check can
fail on Windows because of a mismatch in path separators ('/' vs '\').
This CL fixes the issue by normalizing the path separators in the check.

Change-Id: I5db62be66582cf147f6506a12def6b3911ae0cf9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6429887
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99664}