[compiler-dispatcher] Cancel TF jobs on isolate tear down
This CL allows to cancel Turbofan jobs in-between compilation phases
by the OptimizingCompileDispatcher. For this TF checks a flag with
a relaxed load during compilation and finishes with a failure as soon
as this flag was set.

We cancel jobs on isolate teardown but also whenever we receive
the Isolate::ContextDisposedNotification() from Blink. On isolate
teardown after cancelling we still need to wait until the jobs
finish. On the other hand on context disposal in the non-blocking
variant we just cancel the running jobs but we don't need to wait
until the eventually halt.

This improves the total Turbofan finalization time on isolate tear
down from about 1.5s down to about 0.5s in Jetstream. With that it improves bomb-workers by about +3.7% and bomb-workers.Average by
about 4.3%.

Bug: 408010431
Change-Id: I098454bda187689ea8d9b93e4929837ede482fa0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6404676
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99659}