[arm64][compiler] Omit bounds checks for byte stores
At storing, we skip dynamic bounds checks if PartialOOBWriteAreNops.
This is based on the platform.
This patch skips dynamic bounds check for single byte stores as well,
because those cannot be partially OOB.

This provides a ~4% speedup over a number of wasm benchmarks.

Change-Id: I9e10a2a77fa7e652c1bc864d01dcb8c28d09e008
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6429490
Reviewed-by: Sam Parker-Haynes <sam.parker@arm.com>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Sam Parker-Haynes <sam.parker@arm.com>
Cr-Commit-Position: refs/heads/main@{#99672}