[heap] Add early bailout on allocation slow path for new space
So far it was always the case that we reached
AllocateOnNewPageBeyondCapacity only if the space has already reached
its target capacity. However, now that we lazily allocate new space
pages allocation can also fail when the current capacity is below
the target capacity. E.g. because we run out of virtual memory in
the cage or other OOM situations. This now starts to triggger the
DCHECKs in AllocateOnNewPageBeyondCapacity, which assume that the
space is full. Instead of removing these DCHECKs this CL fixes this
by adding an early bailout in EnsureAllocation(). With this we
only invoke ShouldExpandYoungGenerationOnSlowAllocation() for the
SemiSpaceNewSpace if the space is actually at its capacity.

Bug: 391129951, 408042977
Change-Id: Ib625eec29a2225f7860a9911361feba06bf5bbea
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6432947
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99670}