[js_fuzzer] Implement snippet-DB validation tool
This adds a new script to harden assumptions about snippets in the
DB.

Error-prone snippets often lead to early bail-outs or syntax
errors. This adds a tool that tests the snippets with a test
insertion and actual code execution. We don't yet use the
updated DB index, so this change does not affect production yet.
We'll include update instructions and bootstrapping with the
updated index in a follow up.

This change also removes again the DB validation in the runner
and script mutator and instead reuses some of the code in the new
tools. Previously it was meant to test the quality of the DB, but
it's actually better to use this information directly to improve
its quality.

This adds an end-to-end test that first builds a DB with inputs
with some error-free and some error-prone snippets. The runs the
validation, which skips a snippet with syntax errors and which
marks two error-prone snippets to require try-catch.

Bug: 389069288
Change-Id: Ie377921490cf20c469bb360dec9a7ea40b6412d3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6429728
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#99686}