[js_fuzzer] Use try-catch info from validated DB
This adds instructions to also run the DB validation tool as part
of the build process. The try-catch info from the tool is now used
when retrieving expressions, so that a live analysis is not
necessary. The old analysis code is no longer needed as it wasn't
accurate anyways. Error-prone snippets are now always wrapped with
try catch.

Bug: 389069288
Change-Id: I2137e66d53e26feb525a65a93bb93798d42720e5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6433113
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99687}