[js_fuzzer] Use Fuzzilli inputs for cross-insertions
This adds instructions to also use the Fuzzilli corpus for creating
the cross-insertion snippet DB. Since the number of snippets grows
quite large with this, we also compress the DB slightly by:
- Using smaller hashes for file names (collisions are still very
rare and don't really matter)
- Pruning redundant information from the meta info. We have some
of the information stored in the index config.
- Pruning large snippets.

Because of the higher-quality snippets based on the previous CLs
and better try-catch wrapping, we also increase the probability
of using cross-insertions.

Bug: 362963886
Change-Id: Id1e60dd636bc8cb11364d9486a4c70f48e90cd95
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6433212
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99688}