[wasm] Compilation hints: Fix inconsistent required tiers
in case of --no-wasm-lazy-compilation.

When using eager compilation hints in combination with
--no-wasm-lazy-compilation, the CompilationStateImpl would end up in a
funky state:

GetDefaultTiersPerModule() normally sets the top_tier to
ExecutionTier::kNone but with lazy compilation disabled, it gets set
to the baseline_tier which is initialized with ExecutionTier::kLiftoff.

Afterwards, the ApplyCompilationHintToInitialProgress() can update the
baseline_tier to ExecutionTier::kTurbofan if the hint requests eager
optimized compilation.
The top_tier is not modified by this hint, resulting in the
baseline_tier being Turbofan and the top_tier being Liftoff which
causes a DCHECK failure in case of deopts (for which the ModuleCompiler
needs to AllowAnotherTopTierJob() in case of a liftoff compilation
triggered by a deoptimization.

Bug: 381281321, 42204618
Change-Id: Ib0983b7ebca4426d27151895b44e983752ceadf0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6437350
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99691}