Add condition to check SameValue(O, receiver) for TypedArray [[Set]]
the LookupIterator to evaluate whether the index is found or not.

If the index is found (case LookupIterator::DATA), it already prevents
value coercion via it->HolderIsReceiverOrHiddenPrototype(). However, if
the index is not found (case
LookupIterator::TYPED_ARRAY_INDEX_NOT_FOUND) - it does not
write the value itself, but it incorrectly performs value coercion.

Therefore, add a condition to check via
LookupIterator::HolderIsReceiver() to work the same with index found,
 which is spec [[Set]] intended. [1]

[1] https://tc39.es/ecma262/#sec-typedarray-set

Fixed: 408118033
Change-Id: I7f3a3a2bf26289c338aeac8a251d81eee7a2dd62
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6429205
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Seokho Song <seokho@chromium.org>
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99732}