[wasm][jspi] Reset stack switch info when stack retires
When a stack retires and is moved to the stack pool, clear its
information about temporary switches to the central stack
(stack_switch_info_).
When the stack retires due to a normal return, this field should have
already been cleared when we last switched back from the central stack.
However if the stack retires due to an exception, this field is never
explicitly cleared, and the StackMemory object can be reused for a
different wasm stack in this incorrect state.

R=jkummerow@chromium.org

Fixed: 408254017
Change-Id: I11a66244925d7e59822df52f01c18b055c33891d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6440830
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99727}