[maglev] Fix typed array constant folding
When inlining, the const we're seeing is not necessarily a TypedArray,
or a TypedArray of the right kind, so, we need to check.

This CL reverts back to the simpler but more verbose implementation
where constant typed array loads / stores have their own Maglev nodes.

Bug: 407795724
Change-Id: I0479323142e5825dc454e069454b13797a1222e9
Fixed: 409351652, 409354662
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6439042
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99728}