Add a is_breakable flag to PositionTableEntry
It will be used by the bytecode generator to mark source positions that
are not supposed to be breakable by the debugger.

This doesn't use the flag yet, that will be done as a follow up to mark
mark “iterator.next” calls in for-of bytecode as non-breakable.

See go/chrome-devtools:non-breakable-positions-proposal

Bug: 407773675
Change-Id: I50103aea6a2954cc34ced4d81fb657b885421093
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6438994
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99754}