[compiler] Fix cancellation of TF jobs
Run() did not return the correct failure value when the job got
cancelled. Returning std::nullopt let the job continue to run. As long
as the phase ran successfully this was benign because the job just
ran a bit longer than actually needed. However, this is an issue if
the exact same phase both fails and gets cancelled from a concurrent
thread at the same time. In this case the error value of the phase
would be ignored and the job continues to run.

This CL fixes this by returning BailoutReason::kCancelled instead of
{} when the job got cancelled. This works because we only either
return void or std::optional<BailoutReason> from Run().

Bug: 408010431
Change-Id: I568f89c040b3b024f867971c9c13a3a6aa3f83e1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6443895
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99738}