[maglev] Fix type system refactoring: IsNodeType(empty, Foo) == false
We already check the empty type and bail out more eagerly in CheckString
etc., so this seemsto just work.

It makes it easier to write safer code, like this:
  if (IsNodeType(my_type, kSmi)) {
    unsafe operation assuming it's a Smi
  } else {
    safe operation not assuming it's a Smi
  }

This also fixes a bug where we end up with an impossible type for an
input of a CheckedInternalizedString, and generate unsafe code during
the input conversion (which runs before CheckedInternalizedString would
deopt).

Fixed: 409354670
Change-Id: I8970bc0ad44bba91138b89ff5e0c3d609893404f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6446713
Commit-Queue: Marja Hölttä <marja@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99756}