[wasm][jspi] Disable SP-relative operands if wasm switches stacks
If a function switches to the central stack, this temporarily
invalidates the SP, so we disable using SP-relative memory operands for
this function.
This was only done for wasm builtins and wasm-to-js wrappers so far. We
also need to do this for regular wasm functions now since they may
switch for Fast API calls.
Detect this more precisely based on the presence of the SetStackPointer
instruction.

R=jkummerow@chromium.org

Fixed: 408242499
Change-Id: I6e20c49426fcb83029ae025226959c2636ed95af
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6448591
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99761}