[runtime] Handle termination exception while throwing
In the linked issue, a wasm fuzzer installs a near heap limit callback
which terminates execution, and the callback is called while a wasm
error is being created. So after the callback has returned, we try to
throw the wasm error, but the termination exception has not been
processed and cleared yet and we hit the DCHECK(!has_exception()) in
Isolate::Throw().
Relax the DCHECK and propagate the termination exception instead.

R=jkummerow@chromium.org

Fixed: 409487530
Change-Id: I5a3daea4f6f87bdbba7b9c6c1fff539213a22612
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6448598
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99766}