[wasm] Fix lookup of wrappers marked "is_dying"
We must not add a dying wrapper to the current WasmCodeRefScope.
Doing so creates a possibility that a wrapper that's about to be
deleted by another thread (after going through a Wasm Code GC cycle)
will be deleted again by that WasmCodeRefScope.
To fix this, this patch merges the "is_dying" bit and the refcount
into a single bit field that can be updated atomically, and updates
the WasmImportWrapperCache to use the new, more careful machinery.

This definitely fixes issue 409379692, hopefully also 407003348.

Bug: 407003348
Change-Id: Ib83e9df034e8ddf3590828aea86741779ef92ab9
Fixed: 409379692
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6450127
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99765}