[heap] Support growing of large objects
This CL adds a new method HeapAllocator::TryResizeLargeObject which
allows the user of this API to resize a large page allocation in-place
if possible. This operation may fail if there is not enough free
virtual memory after the end of the page. In such scenarios a user
of this API needs to try to allocate a fresh large object of the
appropriate size instead.

This API will be useful for the JSON stringifier which needs to
use the string as an output buffer which needs to be grown/extended repeatedly.

Note that resizing large objects in this CL only works in builds with
pointer compression. We can support other builds as a follow-up. Until
then TryResizeLargeObject() will just always fail on them.

Bug: 409069381
Change-Id: I80978b14a1d12d8418eff9c293e2162765a14ed1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6438844
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99777}