[maglev] Fix: detect empty type more aggressively when converting inputs
We don't have a way to bail out from AddNewNode, but the inputs to the
node we're creating might already have the empty type.

This CL checks for the empty type more aggressively and ensures we don't
do unsafe operations because of the empty type.

Fixed: 409354670
Change-Id: I196eef97ba7ce0302f0b109628165da67fc05b19
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6439017
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99776}