[arm64][sim] Fix "access below SP" false positive
When an exception is caught in a different stack than the one it was
thrown from, the simulator stack limit and SP register are updated but
are out of sync for a short timespan. This can cause the simulator to
incorrectly interpret some load instructions in the CEntry builtin as
below SP stack reads and crash.
Ensure that we update the simulator stack limit right after we have
updated the SP.

R=jkummerow@chromium.org

Fixed: 408571513
Change-Id: I48f9750cbbf1ebf7e8c2c9151b0ed51b15acebe9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6448908
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99834}