[inspector] Add V8Inspector::connectShared
This CL adds a version of the 'connect' method that returns a shared_ptr
rather than a unique_ptr.

This allows us to defer deconstruction of V8InspectorSession objects.
We use this in "dispatchProtocolMessage" for now.

The main reason is that blink is allowed to tear down V8 sessions on
the nested run-loop, while actual agent/session functions remain on
the stack. This CL attempts to fix this in a more general way than
patching adhocs UaF.

Note that its not clear if this fixes all instances where the session
dies on the nested run loop: There are other entry points into
inspector agent code that can then transition into JS and then run
a nested event loop, e.g. client events or interrupts.

Bug: 40071155
Change-Id: I48f599cfcda2cf8ed71e3b865a6d2e71daa59061
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6433946
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99784}