[json] Stringify: Fix updating DescriptorArray fast iterable state
Only update DescriptorArrays to fast JSON iterable, if we actually
iterated all the properties in the descriptor array.
Due to sharing of DescriptorArrays, we can iterate only a subset of
properties for an object and shouldn't update the state to fast in that
case.

Fixed: 410842391
Bug: 380044242
Change-Id: I9dc273c3f948b70f1d0eedf91ced0f1be9e46b89
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6461722
Auto-Submit: Patrick Thier <pthier@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99806}