[sandbox] Ensure consistency of "current isolate" TLS
There are no known open issues, and we did not see any DCHECK failures
in the last three weeks (see https://crbug.com/405047989).

So this CL turns the DCHECK into a CHECK. This is planned to be the
state for the next branch cut (v13.7) so embedders that do still not
enter the right isolate will get a hard crash instead of
difficult-to-debug random problems from using the wrong pointer tables.

We use a CHECK instead of SBXCHECK to really flush out all bugs in
embedders, even if the sandbox is disabled. In particular
`GetHeapFromWritableObject` is used for things unrelated to the sandbox.

R=ishell@chromium.org CC=​​dinfuehr@chromium.org

Bug: 396607238
Change-Id: Ib16dc2dd8ab71c58b6113299ce910bc03b73b8af
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6458560
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99823}