[parser] Fold string addition on RHS of expression
For patterns like `expr + "abc" + "def"`, the adddition is left
associative, which means this is parsed as `(expr + "abc") + "def"`.
However, string concatenation (unlike floating point addition) is
commutative, so this could be optimized to `expr + ("abc" + "def")`,
which can be constant folded to `expr + "abcdef"`.

We can do this as part of an N-ary operation append. A sequential
addition of a+b+c is transformed from

     +
    / \   + c
   a   b

to

    n-ary +
    /  |  \
   a   b   c

If we see that b and c are both string literals, we can change this
append into an in-place mutation of b, so that we transform

     +
    / \   + "c"
   a  "b"

to

   n-ary +
    /  \
   a   "bc"

Fixed: 411054268
Change-Id: I62c321b2afb9daaf177150804fbbb2f5ea972366
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6459006
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99833}