[json] Stringify: Check if key is fast even if we don't serialize it
We don't serialize property keys if the property value is undefined.
However we should still check if a key is fast iterable in that case, if
we don't know already, to set the correct fast iterable state on the
DescriptorArray.

Fixed: 410979284
Bug: 380044242
Change-Id: Ic25ecfe2cf03856c90942977ed0dca0a5d732999
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6468893
Auto-Submit: Patrick Thier <pthier@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99851}