[string] Cache concatenated strings used as property keys
We special-case string concatenation when the result is used as a
property key. In this case, we know it will eventually be internalized
and it's better to do so early.

For now, we handle only the specialized situation in which lhs is a
string constant, i.e.: `o["foo" + str]`. Here we know that the operation
will be StringAdd (and not an arithmetic addition); and we can use the
rhs as a key into a cache for internalized versions of `"foo" + str`.

Bug: 411350915
Change-Id: Iaa0483786f47b1b5e826e862778f7fcf2c1522df
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6468653
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99824}