[codegen] Add comment for optimizing calls to known JS functions
Currently, when calling a known JSFunction object, we still (often)
execute a SbxCheck at runtime to ensure that the parameter counts match.
Instead of that, we should switch our compilers to generate calls via
the JSDispatchTable with a hardcoded dispatch handle, which avoids the
need for any runtime checks. This CL just some comments that provide a
brief explanation and point to the relevant issue for more details.

Bug: 412398354
Change-Id: I3a77938d06158636ef53e4651656fad1d6d67f24
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6478174
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Auto-Submit: Samuel Groß <saelo@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99840}