[turbofan] Fix --trace-turbo file creation for Wasm C++ template funcs
Currently V8 fails to create the trace-turbo files for Wasm functions
compiled from C++ template functions. This is due to the fact that the
'<' and '>' characters cannot be part of the file name on Windows. (They
do work on Linux.)

For example this is a trace filename for which the file creation fails:
turbo-render(std--__2--vector<Sphere,_std--__2--allocator<Sphere>>_...

As a fix, we can replace the '<' and '>' characters with '{' and '}' in
the trace.

Change-Id: I5e042f6bc3e73d4120dc5197ba8689aacc6ea05c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6478792
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99867}