[wasm][api][source-phase-imports] Remove CodeGenAllowed check from API.
We are using the public API interface `WasmModuleObject::Compile` to
implement source phase imports support in blink. Currently the method
returns an empty object if the `WasmCodeGenAllowed` check fails. As a
consequence, loading wasm modules through ESM is blocked is the
script-src wasm-unsafe-eval CSP is not present.

The only non-test caller of this method in the Chromium codebase is the
auction worklet, and it doesn't set `wasm_code_gen_allowed_callback`
when initializing V8.

Historically, V8 keeps this method for other embedders, see:
https://issues.chromium.org/issues/42200465

Related issue in the WebAssembly esm-integration github:
https://github.com/WebAssembly/esm-integration/issues/77

Bug: 42204365
Change-Id: I46e3fa79822936eeca2cd76b46ad34174391016a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6472866
Commit-Queue: Luis Pardo <lpardosixtos@microsoft.com>
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99845}