[sandbox] Document how SFI::formal_parameter_count can be used safely
As there are some legitimate use cases for SFI::formal_parameter_count,
we will likely not remove the field entirely (though we might be able to
merge it with |length| in the future). Instead, this CL adds some
documentation explaining when the field can be used safely and when one
of the various alternatives should be used instead.

Bug: 40931165
Change-Id: I755e900da7e835f04318bc034184b8bb69609ea8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6480377
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99854}