[wasm] Only use the generic wrapper if enabled
If the --wasm-generic-wrapper flag is disabled, we should not use the
generic wrapper.

This fixes a DCHECK error during wrapper tier up because we did not
update the `call_origin` of `WasmImportData` correctly if that flag was
disabled.

More checks and minor fixes will be added in a follow-up CL.

R=jkummerow@chromium.org

Bug: 410295354
Change-Id: Ibcb54766b055d92721dd9caa69ae1c935dc47d75
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6482475
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99873}