[wasm][jspi] Duplicate string builtins for wasm
Create a copy of the JS string builtins for wasm. Only the wasm version
of the builtins check if they need to switch to the central stack before
calling JS or C++. The JS builtins can assume that they are already
running on the central stack.

R=jkummerow@chromium.org

Bug: 42202153
Change-Id: Iacbda9d2fb1ec243e7b9d3583adb94503ba50b47
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6482073
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99887}