[sandbox] Remove SFI::IsDontAdaptArguments
The SharedFunctionInfo object should not be used to make argument
adaption decisions. Instead, those decisions must be based on the actual
code that will be executed, so must use a parameter count from (for
example) a BytecodeArray or the JSDispatchTable. As such, this CL
removes the IsDontAdaptArguments accessor from the SFI.

Bug: 40931165
Change-Id: I9a4634897f6619692c369aab570f8ee117794624
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6485688
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99892}