[sandbox] Mark SFI::formal_paramater_count as deprecated in heap-refs
In the compiler we should not need to look at the parameter count on the
SFI and instead use a (trusted) parameter count, for example from the
BytecodeArray or JSDispatchTable. This CL therefore marks the respective
fields on the HeapBroker as deprecated and fully removes the
IsDontAdaptArguments accessor as that is already unused.

Drive-By: Introduce FrameStateInfo::parameter_count_without_receiver()
for convenience.
Drive-By: Make some FrameStateInfo getters more robust so they fail a
DCHECK instead of returning a value when the data is unavailable.

Bug: 370343328
Change-Id: I3ffec15817db4178b614ba67235b831679f86015
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6481263
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99889}